requirements.txt
# Real-Time Intelligent Surveillance System - Dependencies
# Install with: pip install -r requirements.txt

# ============================================================================
# WEEK 1-2: CORE DEPENDENCIES (Required)
# ============================================================================

# OpenCV for video processing and visualization
opencv-python>=4.8.0

# Numerical computing
numpy>=1.24.0

# Configuration management
pyyaml>=6.0

# YOLOv8 from Ultralytics (Week 1: Detection)
ultralytics>=8.0.200

# PyTorch (required by ultralytics)
torch>=2.0.0
torchvision>=0.15.0

# Week 2: Tracking (DeepSORT)
deep-sort-realtime>=1.3.0

# ============================================================================
# FUTURE WEEKS (Optional for now)
# ============================================================================

# Week 3: Dashboard
streamlit>=1.28.0
pandas>=2.0.0
matplotlib>=3.7.0

# ============================================================================
# GPU SETUP NOTES
# ============================================================================

# For NVIDIA GPU acceleration, install CUDA-enabled PyTorch:
# CUDA 11.8: pip install torch torchvision --index-url https://download.pytorch.org/whl/cu118
# CUDA 12.1: pip install torch torchvision --index-url https://download.pytorch.org/whl/cu121

# Verify GPU is available:
# python -c "import torch; print(torch.cuda.is_available())"
